wip: initial rewrite on top of async/await APIs
The general approach of using an Engine which drives the "IRC thread", as
well as an asynchronous channel for communicating with the Engine seems to
be working well. The testing wrapper is allowing for a really fluent testing
API (which will only be improved as more progress is made).

Currently, linking is implemented (each Link is a separate Future currently,
the idea being that each one can be tokio::spawn'd independently. In the
future (haha) this could be migrated to something like the FuturesUnordered
API. Servers and subnets are bursted upon linking, but clients are currently
not shared.

In addition, the test suite consists only of one example test, and needs to
be expanded signficantly.