wip: multi-server split test working
This commit adds a test for an a <-> b <-> c <-> d linking scenario
where the link c <-> d is broken.

To make this work, broadcasting of bursting and splitting was
implemented.

Additionally, a bug was fixed where within the TestEngine, a sync()
request would get stuck if syncing triggered processing of netsplit
messages. This happened because sync() took a snapshot of the count
of connected servers, and due to the split, not all those connected
would respond to the sync. Instead, the Netsplit() event is now the
indicator that a subset of servers should be considered synced even
without a SyncResponse.